+++ README.md
@@ -1,5 +1,5 @@
 # mesto 
-## [сайт Mesto v1.3 ](https://ordanan.github.io/mesto/)
+## [сайт Mesto v1.6 ](https://ordanan.github.io/mesto/)
 
 #### В данной работе, мной был сделан блок на Java Script:
 
@@ -13,4 +13,10 @@
 - удаление своей карточки
 - возможность обновлять аватар пользователя
 
-Сборка проекта осуществлялась Webpack.  
+Сборка проекта осуществлялась Webpack.
+//в файле readme.md отсутсвует информация о том как развернуть проект. Необходимо указать какие виды сборок присутствуют в проекте, как их запустить
+Сборка в режиме dev (сборка для разработки) используется локальный сервер. Для запуска сборки, вызовите из терминала команду npm run dev. Проект откроется на локальном сервере — по адресу localhost:8080.
+
+Сборка в режиме build (production сборка). Вызывается из терминала командой npm run build. В результате сборки создается директория dist. В этой директории фаил main.js будет содержать весь JS-код проекта.
+
+Сборка в режиме deploy. Соберите ваш проект командой npm run deploy. Вебпак скопирует собранный проект из папки dist в ветку gh-pages. После этого вы можете посмотреть проект в интернете по адресу https://ordanan.github.io/mesto/ .
